chore(components/actions): move action.ts under components
